refactor(movements): Move existing accounts to movements namespace
